@@ -5,7 +5,7 @@ description: >-
   as it animates to the new route.
 
 environment:
-  sdk: ^3.5.0
+  sdk: ^3.6.0-0
 
 dependencies:
   flutter:
