@@ -5,7 +5,7 @@ description: >-
 version: 1.0.0
 
 environment:
-  sdk: ^3.5.0
+  sdk: ^3.6.0-0
 
 dependencies:
   flutter:
