@@ -25,7 +25,7 @@ class Photo extends StatelessWidget {
   Widget build(BuildContext context) {
     return Material(
       // Slightly opaque color appears where the image has transparency.
-      color: Theme.of(context).primaryColor.withOpacity(0.25),
+      color: Theme.of(context).primaryColor.withValues(alpha: 0.25),
       child: InkWell(
         onTap: onTap,
         child: LayoutBuilder(
