@@ -61,7 +61,7 @@ class _ExampleInstagramFilterSelectionState
         return Image.network(
           'https://docs.flutter.dev/cookbook/img-files'
           '/effects/instagram-buttons/millennial-dude.jpg',
-          color: color.withOpacity(0.5),
+          color: color.withValues(alpha: 0.5),
           colorBlendMode: BlendMode.color,
           fit: BoxFit.cover,
         );
