@@ -122,7 +122,7 @@ class FilterItem extends StatelessWidget {
             child: Image.network(
               'https://docs.flutter.dev/cookbook/img-files'
               '/effects/instagram-buttons/millennial-texture.jpg',
-              color: color.withOpacity(0.5),
+              color: color.withValues(alpha: 0.5),
               colorBlendMode: BlendMode.hardLight,
             ),
           ),
